requirements/requirements.txt
discord.py
requests
lxml
pytz
python_weather
pyyaml
